requirements.txt
joblib>=0.14
lazy-loader>=0.1
msgpack>=1.0
numba>=0.51.0
scikit-learn>=0.20.0
scipy>=1.2.0
streamlit>=1.3.1
customtkinter==5.2.2
dashscope==1.19.2
fuzzywuzzy==0.18.0
logging_reset==10.0.3
numpy==1.26.4
paddleocr==2.7.3
Pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyperclip==1.8.2
Requests==2.32.3
SpeechRecognition==3.10.4
vosk==0.3.45
websocket_client==1.8.0
joblib>=0.14
lazy-loader>=0.1
msgpack>=1.0
numba>=0.51.0
scikit-learn>=0.20.0
scipy>=1.2.0
streamlit>=1.3.1
PyAudio==0.2.14
paddlepaddle
python-Levenshtein==0.25.1